+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+ * Fancy text formatting renderer
+
 ## 0.6.10 - 2021-09-18
  * Glaxnimate compatibility